+++ packages/flow/src/utils/edge/get-edges-by-node.ts
@@ -1,0 +1,9 @@
+import { FlowEdge } from '../../interfaces/edge';
+import { FlowElementType } from '../../interfaces/element';
+import { PlaitBoard } from '@plait/core';
+import { getFlowElementsByType } from '../get-node-by-id';
+
+export const getEdgesByNodeId = (board: PlaitBoard, nodeId: string) => {
+    const edges = getFlowElementsByType(board, FlowElementType.edge) as FlowEdge[];
+    return edges.filter(item => item.target.id === nodeId || item.source?.id === nodeId);
+};