+++ packages/plait/src/plugins/with-selection.ts
@@ -4,11 +4,18 @@ import { Transforms } from '../transforms';
 import { transformPoint } from '../utils/board';
 import { toPoint } from '../utils/dom';
 import { RectangleClient } from '../interfaces/rectangle-client';
-import { cacheSelectedElements, calcElementIntersectionSelection } from '../utils/selected-element';
+import {
+    cacheSelectedElements,
+    calcElementIntersectionSelection,
+    getSelectedElements,
+    isIntersectionElements
+} from '../utils/selected-element';
 import { SELECTION_BORDER_COLOR, SELECTION_FILL_COLOR } from '../interfaces';
 import { getRectangleByElements } from '../utils/element';
+import { BOARD_TO_IS_SELECTION_MOVING, BOARD_TO_TEMPORARY_ELEMENTS } from '../utils/weak-maps';
+import { ATTACHED_ELEMENT_CLASS_NAME } from '../constants/selection';
 
-export function withSelection<T extends PlaitBoard>(board: T) {
+export function withSelection(board: PlaitBoard) {
     const { mousedown, globalMousemove, globalMouseup, onChange } = board;
 
     let start: Point | null = null;
@@ -17,13 +24,23 @@ export function withSelection<T extends PlaitBoard>(board: T) {
     let selectionOuterG: SVGGElement;
 
     board.mousedown = (event: MouseEvent) => {
-        selectionOuterG?.remove();
         if (event.button === 0) {
             start = transformPoint(board, toPoint(event.x, event.y, PlaitBoard.getHost(board)));
         }
         if (start) {
-            Transforms.setSelection(board, { ranges: [{ anchor: start, focus: start }] });
+            const ranges = [{ anchor: start, focus: start }];
+            const selectedElements = getSelectedElements(board);
+            const intersectionSelectedElement = isIntersectionElements(board, selectedElements, ranges);
+            if (intersectionSelectedElement) {
+                start = null;
+            } else {
+                Transforms.setSelection(board, { ranges: ranges });
+                if (calcElementIntersectionSelection(board).length) {
+                    start = null;
+                }
+            }
         }
+
         mousedown(event);
     };
 
@@ -31,13 +48,11 @@ export function withSelection<T extends PlaitBoard>(board: T) {
         if (start) {
             const movedTarget = transformPoint(board, toPoint(event.x, event.y, PlaitBoard.getHost(board)));
             const { x, y, width, height } = RectangleClient.toRectangleClient([start, movedTarget]);
+            selectionMovingG?.remove();
             if (Math.hypot(width, height) > 5) {
                 end = movedTarget;
-                if (end) {
-                    Transforms.setSelection(board, { ranges: [{ anchor: start, focus: end }] });
-                }
-                PlaitBoard.getBoardNativeElement(board).classList.add('selection-moving');
-                selectionMovingG?.remove();
+                Transforms.setSelection(board, { ranges: [{ anchor: start, focus: end }] });
+                setSelectionMoving(board);
                 const rough = PlaitBoard.getRoughSVG(board);
                 selectionMovingG = rough.rectangle(x, y, width, height, {
                     stroke: SELECTION_BORDER_COLOR,
@@ -53,8 +68,18 @@ export function withSelection<T extends PlaitBoard>(board: T) {
 
     board.globalMouseup = (event: MouseEvent) => {
         if (start && end) {
-            PlaitBoard.getBoardNativeElement(board).classList.remove('selection-moving');
             selectionMovingG?.remove();
+            clearSelectionMoving(board);
+        }
+
+        if (PlaitBoard.isFocus(board)) {
+            const isInBoard = event.target instanceof Node && PlaitBoard.getBoardNativeElement(board).contains(event.target);
+            const isAttachedElement = event.target instanceof HTMLElement && event.target.closest(`.${ATTACHED_ELEMENT_CLASS_NAME}`);
+            // Clear selection when mouse board outside area
+            // The framework needs to determine whether the board is focused through selection
+            if (!isInBoard && !start && !isAttachedElement) {
+                Transforms.setSelection(board, null);
+            }
         }
 
         start = null;
@@ -65,13 +90,14 @@ export function withSelection<T extends PlaitBoard>(board: T) {
     board.onChange = () => {
         // calc selected elements entry
         try {
+            selectionOuterG?.remove();
             if (board.operations.find(value => value.type === 'set_selection')) {
-                const elementIds = calcElementIntersectionSelection(board);
-                cacheSelectedElements(board, elementIds);
-                const { x, y, width, height } = getRectangleByElements(board, elementIds, false);
+                const temporaryElements = getTemporaryElements(board);
+                const elements = temporaryElements ? temporaryElements : calcElementIntersectionSelection(board);
+                cacheSelectedElements(board, elements);
+                const { x, y, width, height } = getRectangleByElements(board, elements, false);
                 if (width > 0 && height > 0) {
                     const rough = PlaitBoard.getRoughSVG(board);
-                    selectionOuterG?.remove();
                     selectionOuterG = rough.rectangle(x - 2, y - 2, width + 4, height + 4, {
                         stroke: SELECTION_BORDER_COLOR,
                         strokeWidth: 1,
@@ -80,6 +106,7 @@ export function withSelection<T extends PlaitBoard>(board: T) {
                     PlaitBoard.getHost(board).append(selectionOuterG);
                 }
             }
+            deleteTemporaryElements(board);
         } catch (error) {
             console.error(error);
         }
@@ -87,4 +114,26 @@ export function withSelection<T extends PlaitBoard>(board: T) {
     };
 
     return board;
+}
+
+export function getTemporaryElements(board: PlaitBoard) {
+    return BOARD_TO_TEMPORARY_ELEMENTS.get(board);
+}
+
+export function deleteTemporaryElements(board: PlaitBoard) {
+    BOARD_TO_TEMPORARY_ELEMENTS.delete(board);
+}
+
+export function isSelectionMoving(board: PlaitBoard) {
+    return !!BOARD_TO_IS_SELECTION_MOVING.get(board);
+}
+
+export function setSelectionMoving(board: PlaitBoard) {
+    PlaitBoard.getBoardNativeElement(board).classList.add('selection-moving');
+    BOARD_TO_IS_SELECTION_MOVING.set(board, true);
+}
+
+export function clearSelectionMoving(board: PlaitBoard) {
+    PlaitBoard.getBoardNativeElement(board).classList.remove('selection-moving');
+    BOARD_TO_IS_SELECTION_MOVING.delete(board);
 }