+++ packages/plait/src/public-api.ts
@@ -13,4 +13,5 @@ export * from './interfaces';
 export * from './plait.module';
 export * from './transforms';
 export * from './utils';
-
+export * from './plugins/with-selection';
+export * from './plugins/with-moving';