+++ packages/plait/src/utils/board.ts
@@ -1,5 +1,6 @@
 import { PlaitBoard } from '../interfaces/board';
 import { Point } from '../interfaces/point';
+import { distanceBetweenPointAndRectangle } from './math';
 
 export type ViewBox = {
     minX: number;
@@ -27,6 +28,9 @@ export function transformPoint(board: PlaitBoard, point: Point) {
     return newPoint;
 }
 
-export function getBoardViewportContainer(board: PlaitBoard) {
-    return PlaitBoard.getHost(board).parentElement as HTMLElement;
+export function isInPlaitBoard(board: PlaitBoard, x: number, y: number) {
+    const plaitBoardElement = PlaitBoard.getBoardNativeElement(board);
+    const plaitBoardRect = plaitBoardElement.getBoundingClientRect();
+    const distances = distanceBetweenPointAndRectangle(x, y, plaitBoardRect);
+    return distances === 0;
 }